+++ app/components/items/index.js
@@ -1,1 +1,2 @@
+export { default as ValArrayItem } from './ValArrayItem';
 export { default as ValItem } from './ValItem';